adding test for missing attribute name (#87)
If a Section does not have a `name` we can not lint further. This is the case because we need the `name` to match up the plugin with the allowed properties. 

The linter will annotate where the name is missing. 

I've also done some minor refactoring on messages to avoid duplication.