ignore binding error (#98)
The majority of options are easy to run without a cloud context. But some parameters need more context. This is the case of the "Listen" parameter, which as you might suspect, it requires something to listen to. We will ignore annotating this type of option due to the lack of context in the cloud setup.